Update space.computer.json
all 3 languages seem to have the same text here, but I can verify in English, it is not rendering the proper text, but rather the identifier string. I don't remember enough about text in Java or why this was happening, so I just made it match the other two lines that did show correctly.