Attempt at a Russian Item File
I have only taken 2 semesters of Russian, so I cannot promise any high quality of translation here, but I did extensive research on words and proper terms for Minecraft, etc.
There are still some words missing; but I'm hoping that if it gets started, then collaborators may help fill in missing terms, etc.
I just thought it might be interesting given the population size of Russian players.

Feel free to not commit it. I'm making it a separate branch in case you think this might not be a good idea. (Don't worry if this is the case, I still enjoyed learning a ton of new vocabulary!)